requirements.txt
altair==6.0.0
annotated-doc==0.0.4
anyio==4.12.1
attrs==25.4.0
audioread==3.1.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
exceptiongroup==1.3.1
filelock==3.20.0
Flask==3.1.3
fonttools==4.61.1
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.3.1
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
imbalanced-learn==0.14.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
narwhals==2.17.0
networkx==3.4.2
numba==0.64.0
numpy==2.2.6
opencv-python==4.13.0.92
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049101700/work
pandas==2.3.3
pillow==12.0.0
platformdirs==4.9.2
pooch==1.9.0
protobuf==6.33.5
pyarrow==23.0.1
pycparser==3.0
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.5
smmap==5.0.2
sounddevice==0.5.5
soundfile==0.13.1
soxr==1.0.0
streamlit==1.54.0
streamlit-audiorecorder==0.0.6
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0+cpu
torchaudio==2.10.0+cpu
torchvision==0.25.0+cpu
tornado==6.5.4
tqdm==4.67.3
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
Werkzeug==3.1.6
